chore(services): Split services per their responsibilities after the project vision has been changing recently
