+++ trade_portal/devops/localdocker/common.env
@@ -28,6 +28,7 @@ OA_WRAP_API_URL=http://docker-host:9010
 OA_UNPROCESSED_QUEUE_URL=
 OA_UNPROCESSED_BUCKET_NAME=
 OA_AWS_ACCESS_KEYS=:
+AWS_REGION=ap-southeast-2
 OA_VERIFY_API_URL=https://openattverify.c1.devnet.trustbridge.io/verify/fragments
 UA_BASE_HOST=https://trade.c1.devnet.trustbridge.io/v/
 OA_RENDERER_HOST=https://renderer-openatt.c2.devnet.trustbridge.io
