+++ trade_portal/trade_portal/documents/tests/tests_services_lodge.py
@@ -2,8 +2,6 @@ import json
 from unittest import mock
 
 import pytest
-from django.conf import settings
-
 from trade_portal.documents.models import Document, DocumentHistoryItem
 from trade_portal.documents.services.lodge import DocumentService
 from trade_portal.documents.tests.factories import DocumentFactory
@@ -21,7 +19,7 @@ class MockResponse(mock.MagicMock):
 @pytest.mark.django_db
 @pytest.mark.parametrize("ERROR_CODE", [None, "wrap_500", "fail_notarize"])
 @mock.patch("trade_portal.documents.tasks.document_oa_verify.apply_async")
-@mock.patch("trade_portal.documents.services.lodge.NotaryService")
+@mock.patch("trade_portal.documents.services.notarize.NotaryService.notarize_file")
 def test_document_service(notary_service_mock, oa_task_mock, docapi_env, ERROR_CODE):
     mockedJsonResp = {
         "1": "2",
@@ -36,7 +34,7 @@ def test_document_service(notary_service_mock, oa_task_mock, docapi_env, ERROR_C
         status_code=500 if ERROR_CODE == "wrap_500" else 200,
         json_resp=mockedJsonResp if ERROR_CODE != "wrap_500" else {}
     )
-    notary_service_mock.notarize_file.return_value = True if ERROR_CODE != "fail_notarize" else False
+    notary_service_mock.return_value = True if ERROR_CODE != "fail_notarize" else False
 
     s = DocumentService(ig_client=ig_client, oa_client=oa_client)
     doc = DocumentFactory(
@@ -54,13 +52,8 @@ def test_document_service(notary_service_mock, oa_task_mock, docapi_env, ERROR_C
         assert doc.verification_status == Document.V_STATUS_PENDING
         assert doc.workflow_status == Document.WORKFLOW_STATUS_ISSUED
 
-        assert notary_service_mock.notarize_file.call_count == 1
-        notary_service_mock.notarize_file.assert_called_once_with(
-            "{}.{}.{}".format(
-                settings.ICL_APP_COUNTRY.upper(),
-                (doc.created_by_org.business_id).replace(".", "-"),
-                doc.short_id,
-            ),
+        assert notary_service_mock.call_count == 1
+        notary_service_mock.assert_called_once_with(
             json.dumps(mockedJsonResp)
         )
         assert oa_task_mock.call_count == 1
@@ -75,12 +68,12 @@ def test_document_service(notary_service_mock, oa_task_mock, docapi_env, ERROR_C
         assert doc.status == Document.STATUS_FAILED
         assert doc.verification_status == Document.V_STATUS_ERROR
         assert doc.workflow_status == Document.WORKFLOW_STATUS_NOT_ISSUED
-        assert notary_service_mock.notarize_file.call_count == 0
+        assert notary_service_mock.call_count == 0
         assert oa_client.wrap_document.call_count == 1
         assert oa_task_mock.call_count == 0
     elif ERROR_CODE == "fail_notarize":
         assert issue_result is True
-        assert notary_service_mock.notarize_file.call_count == 1
+        assert notary_service_mock.call_count == 1
         assert oa_client.wrap_document.call_count == 1
         assert oa_task_mock.call_count == 0
         assert (doc.status, doc.verification_status, doc.workflow_status) == (
