+++ trade_portal/trade_portal/websub_receiver/tasks.py
@@ -1,6 +1,6 @@
 import logging
 
-from trade_portal.documents.services.lodge import NodeService
+from trade_portal.documents.services.igl import IGLService
 from config import celery_app as app
 
 logger = logging.getLogger(__name__)
@@ -9,6 +9,6 @@ logger = logging.getLogger(__name__)
 @app.task(ignore_result=True, max_retries=3)
 def subscribe_to_new_messages():
     try:
-        NodeService().subscribe_to_new_messages()
+        IGLService().subscribe_to_new_messages()
     except Exception as e:
         logger.exception(e)
